Move `VariantContext` into `:languageNative` test fixture
