+++ subprojects/ide-native/src/integTest/groovy/org/gradle/ide/visualstudio/VisualStudioCppApplicationProjectIntegationTest.groovy
@@ -16,6 +16,7 @@
 
 package org.gradle.ide.visualstudio
 
+import org.gradle.language.VariantContext
 import org.gradle.nativeplatform.fixtures.app.CppApp
 import org.gradle.nativeplatform.fixtures.app.CppSourceElement
